Check if i2c reply is 31DA
On non-CVE systems (WPU) there can be considerable i2c traffic.
After a 31DA request, we cannot be certain that the next message
is a 31DA reply. (On non-CVE systems it will never be a 31DA)
Check if reply is 31DA to prevent incorrectly reading a random
message as 31DA content.