Expose when the last RF command was received
The API call to get the status of all RF remotes now includes a
timestamp of the last received command. This makes it easier for custom
software to detect changes, and makes it possible to implement custom
timers (which requires knowing how long it has been since the last
command).

See https://github.com/arjenhiemstra/ithowifi/issues/157 for more
details.