Add API support for managing settings
This adds the ability to get and set settings using the API. Settings
are obtained using the "getsetting" parameter, while updating settings
is done using "setsetting" (for the index) and "value" (for the value).

All settings that are supported can be changed through the API, instead
of only allowing a limited subset. Due to the different settings
supported across devices and firmware versions maintaining an allow/deny
list isn't practical. It's also not clear what settings should or
shouldn't be allowed in the first place.

This fixes #114